Add two flaky tests identified in PR #30455
During that PR, E2E tests were ran twice against commit df561f9. In the
first run, the only test that failed was 'action settings are disabled
for non-channel admin'; in the second run, the only test that failed was
'MM-T5671 should handle mention counts correctly when marking a thread
as unread and unfollowing it'. Thus, I'm marking both as flaky here.

The PR was:
- https://github.com/mattermost/mattermost/pull/30455

The offending cycles were:
- https://automation-dashboard.vercel.app/cycle/13859513327_1-df561f9-pr-onprem-ent
- https://automation-dashboard.vercel.app/cycle/13861853096_1-df561f9-pr-onprem-ent